Add a barebone composer.json and use the composer autoloader
