+++ composer.json
@@ -1,0 +1,17 @@
+{
+    "name": "woocommerce/openai-product-feed-for-woo",
+    "description": "OpenAI Product Feed for WooCommerce",
+    "homepage": "https://github.com/woocommerce/OpenAI-Product-Feed",
+    "type": "wordpress-plugin",
+    "license": "GPL-3.0-or-later",
+    "prefer-stable": true,
+    "minimum-stability": "dev",
+    "require": {
+        "php": ">=7.4"
+    },
+    "autoload": {
+        "psr-4": {
+            "OAPFW\\": "includes"
+        }
+    }
+}
