+++ includes/Core/Interfaces/SerializerInterface.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OAPFW\Core\Interfaces;
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
+/**
+ * Serializer interface
+ */
+interface SerializerInterface {
+
+	public function serialize( array $data ): string;
+	public function getContentType(): string;
+	public function getFileExtension(): string;
+}
