+++ includes/Platforms/OpenAI/Mappers/ProductMapper.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace OAPFW\Platforms\OpenAI\Mappers;
 
-use OAPFW\Core\ProductMapperInterface;
-use OAPFW\Core\SettingsRepositoryInterface;
+use OAPFW\Core\Interfaces\ProductMapperInterface;
+use OAPFW\Core\Interfaces\SettingsRepositoryInterface;
 use OAPFW\Utils\StringHelper;
 
 if ( ! defined( 'ABSPATH' ) ) {
