+++ includes/Settings/SettingsRepository.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace OAPFW\Settings;
 
-use OAPFW\Core\SettingsRepositoryInterface;
+use OAPFW\Core\Interfaces\SettingsRepositoryInterface;
 use OAPFW\Utils\StringHelper;
 
 if ( ! defined( 'ABSPATH' ) ) {
